Updated README.md, renamed ThemeProvider folder and files
