--- src/components/Theme-provider/index.js
@@ -1,3 +1,0 @@
-import ThemeProvider from "./Theme-provider";
-
-export default ThemeProvider;
