+++ src/components/ThemeProvider/ThemeProvider.js
@@ -1,0 +1,22 @@
+import React from "react";
+import PropTypes from "prop-types";
+import { ThemeProvider as Provider } from "styled-components";
+import GlobalStyle from "../../utils/gloabal-styles";
+
+const ThemeProvider = props => {
+  const { theme, children } = props;
+
+  return (
+    <Provider theme={theme}>
+      <GlobalStyle />
+      {children}
+    </Provider>
+  );
+};
+
+ThemeProvider.propTypes = {
+  children: PropTypes.any,
+  theme: PropTypes.object
+};
+
+export default ThemeProvider;
