+++ src/components/ThemeProvider/index.js
@@ -1,0 +1,3 @@
+import ThemeProvider from "./ThemeProvider";
+
+export default ThemeProvider;
