+++ src/index.js
@@ -1,4 +1,4 @@
-export { default as Text } from './components/Text'
-export { default as Box } from './components/Box'
-export { default as ThemeProvider } from './components/Theme-provider'
-export * as Themes from './themes'
+export { default as Text } from "./components/Text";
+export { default as Box } from "./components/Box";
+export { default as ThemeProvider } from "./components/ThemeProvider";
+export * as Themes from "./themes";
